mitkarte/requirements.txt
appdirs==1.4.4
autopep8==1.5.6
black==20.8b1
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.1.0
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.8.1
psycopg2-binary==2.8.6
pycodestyle==2.7.0
regex==2020.11.13
SQLAlchemy==1.4.11
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
WTForms==2.3.3
